fix(tags): reorg some files and folders for Tags
